requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
decorator==4.4.0
Django==2.2.1
django-cors-headers==3.0.2
django-rest-swagger==2.2.0
djangorestframework==3.9.3
Faker==0.9.1
idna==2.8
importlib-metadata==0.18
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
MarkupSafe==1.1.1
mixer==6.1.3
more-itertools==7.0.0
openapi-codec==1.3.2
packaging==19.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.12.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.0
pytest==4.6.3
pytest-cov==2.7.1
pytest-django==3.5.0
#pytest-ipdb==0.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.7
zipp==0.5.1
